fix(ktoggle): ts improvements for ktoggle
