+++ src/types/index.ts
@@ -14,8 +14,8 @@ export * from './dropdown'
 export * from './empty-state'
 export * from './external-link'
 export * from './file-upload'
-export * from './input'
 export * from './input-switch'
+export * from './input'
 export * from './label'
 export * from './method-badge'
 export * from './modal'
@@ -27,14 +27,15 @@ export * from './radio'
 export * from './segmented-control'
 export * from './select'
 export * from './skeleton'
+export * from './slider'
 export * from './stepper'
 export * from './swrv'
 export * from './table'
 export * from './tabs'
 export * from './text-area'
 export * from './toaster'
+export * from './toggle'
 export * from './tooltip'
 export * from './tree-list'
 export * from './truncate'
 export * from './utils'
-export * from './slider'
